Add support for creating self-contained recording playback packages
This adds a build option REACT_APP_NO_ROUTER that removes the code to
read the recordId from the URL. When used in combination with
PUBLIC_URL=. it will load all resources from the current directory. With
this setup, you can put the playback files and recording media files
together into a single directory for a self-contained recording.