Build base for single content playback
This work is a starting point for building a more dynamic and automatic layout
for the recording's playback.

This first idea is to provide a better space usage for recordings that do
not provide presentation related data, such as slides, annotations, cursor
and panzooms nor screenshare (which occupies the same location at the playback).

To use this option in custom builds and avoid the playback trying to fetch for
files that will never exist in a recording, the missing files can be removed
from files.data at config.json.

e.g. for a webcams with closed captions and chat case:
...
  "files": {
    "data": {
      "captions": "captions.json",
      "chat": "slides_new.xml",
      "metadata": "metadata.xml"
    },
...