Fetch shared notes HTML exported file
Etherpad is able to export an HTML version of the meeting's shared notes. We
already store this data at the archive stage of the recordings flow and all we
have to do to enable the notes at the playback is to copy this notes.html file
to the meeting's published directory root.