feat(config): disable the recording date
Add option to hide the recording date from the playback UI.

IMPORTANT: this does not remove user's complete access to the date the
recording was made. BigBlueButton uses the epoch timestamp as part
of the `recordingId`. Anyone a little more curious can still get that
number and convert it to a human readable date.