Fix loading slide images from a relative directory in tldraw v2
The tldraw v2 code was prepending the origin of the current page to
the slide image urls, if they didn't start with "http". This resulted
in it generating invalid URLs that look like:
https://example.compresentation/ea5b16cd4eef022e866b1462aed02a43342a80a8-1726509771537/svgs/slide1.svg
when the playback is compiled to load from a relative URL (i.e. with
`REACT_APP_NO_ROUTER=1 PUBLIC_URL=.` )

The cause appears to be that some dead code (assignment to an unused
variable) was copied from the tldraw v1 code, and then code later in the
file was adjusted to use that variable, making its behaviour differ from
tldraw v1.

Remove the unused code from tldraw v1, and adjust tldraw v2 to match the
behaviour of tldraw v1.